requirements.txt
pytest
playwright
pytest-playwright
pytest-xdist
allure-pytest
requests
qase-pytest
python-dotenv
